Stopped user from making any changes to transformation until they selected a landuse. Also fixed bugs with the new phos calcs for transformations
